separate out structure of getting a graph and updating the graph
`searchRegex` is now removed as when it is set, it really just sets the `filterDatas`.

when looking at filteredSteps, depends on reactIds will already be a part of the final filtered graph. No need to dynamically search.  Same with define or updateNodeLabel.

Move logic of idle skipping to the idle step.  This prevents another pass over the visible filtered steps.

The goal is to update finalFilteredGraph whenever filterDatas updates.  SearchRegex will update filtered data directly.  Getting a filtered graph at step k should never be setting any info on filteredDatas or finalFilteredGraph (it was before).