Enhance movement within a subgraph to use a consistent list (#13)
* add placeholder text to search bar

* fix invalid html div

* force a redraw when window size is changed

debounce every 250ms but allow a redraw every 1 sec

* all graph locations are union'ed, sort'ed, and unique'd by lodash

* fix bugwhere stepsOutputCal should be stepsIdle. bad copy pasta

* when filtering, calculate all valid steps for solid consistency.

added a helper method to use for calculating the next step given an array of steps.

* if there is no previous step, go to the first step (or next/last step)

* use the final graph to do filtering and then subset graph at step i from the final filtered graph

* remove major comment blocks

* if the last idle position can't be found, use the last log length step

* remove unused var

* move graph resize method to it's own file. add resize method to 'f' key

* fix missing method for flow

* Make flow happy

* Add comments use better variables

* Do not include the labels when making a graph.

Instead, use a space of 400 units. and a label up to 350 units.
Use 10k units when hovering.  This should never be reached.

* add a legend value of "Not Yet Defined" and set background opacity to 1

* Set cytoLabel to cytoLabel_ as cytoscape doesn't catch a difference in a getter method

* display the calc time if displaying node time is allowed

* remove undefined legend item

* consistently use maxTextWidth for cytostyle

* allow a graph to be created with `new Graph(oldGraph)`

This applies to Node, Edge, GhostEdge, ActiveStateStatus, HoverStatus, and StatusArr

* set node label to `label (time)` from `label; time`

* separate out structure of getting a graph and updating the graph

`searchRegex` is now removed as when it is set, it really just sets the `filterDatas`.

when looking at filteredSteps, depends on reactIds will already be a part of the final filtered graph. No need to dynamically search.  Same with define or updateNodeLabel.

Move logic of idle skipping to the idle step.  This prevents another pass over the visible filtered steps.

The goal is to update finalFilteredGraph whenever filterDatas updates.  SearchRegex will update filtered data directly.  Getting a filtered graph at step k should never be setting any info on filteredDatas or finalFilteredGraph (it was before).

* when dblclick'ing, set the search bar info to the key of the element

GhostEdge, Node - key
Edge - ghostKey

* bump version